+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/serializer/PackageStateSerializer.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2009--2015 Red Hat, Inc.
+ * Copyright (c) 2020 SUSE LLC
  *
  * This software is licensed to you under the GNU General Public License,
  * version 2 (GPLv2). There is NO WARRANTY for this software, express or
@@ -22,7 +22,6 @@ import redstone.xmlrpc.XmlRpcSerializer;
 
 import java.io.IOException;
 import java.io.Writer;
-import java.util.Date;
 
 /**
  *
